requirements.txt
chromadb==0.4.24
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
sentence-transformers==2.2.2